ATR-855: refactored out event related utils into a separate helper
